test: ensure BTree inserts elements correct
